streaming/requirements.txt
pyspark==3.5.1
redis==5.0.4
python-dotenv==1.0.1
